Split Target.Recalculate into methods with and without invalidation.
TargetExtensions is moved into its own file.
